refactor personal info form
