+++ src/pages/conference/people/staff/detail/changePassword/style.scss
@@ -7,7 +7,7 @@
 .changePass {
   text-align: left !important;
 }
-.firstColunm {
+.colunm {
   height: 72px !important;
   align-items: center;
   &.old {
