+++ src/pages/conference/people/staff/detail/personalInfo/style.scss
@@ -23,7 +23,16 @@
   font-size: 14px !important;
   height: 44px !important;
   &.subname{
-    width: 120px !important;
     margin-right: 10px;
   }
+}
+.birthday{
+  width: 100% !important;
+}
+.gender{
+  border: none;
+  border-bottom: solid 1px !important;
+  width: 100%;
+  height: 24px;
+  margin-top: 12px;
 }
